FERMO Version 0.8.6: Bugfix in quantitative biological data parsing
Fixes an error which occured when the value in the 'quant_data' column
of the first row (excluding the header row) was 0. Rows with value 0
are filtered from the table using Pandas. This creates gaps in the
index, which led to errors in the Pandas engine.

Bugfix: Reset index of the filtered dataframe. This fixed the
error.