0.8.8.1: Change to fold-change filter on dashboard page
Added additional functionality to fold-change filter.
Instead of only a filter for features with fold-changes greater or
equal to a certain integer value, the group that should be searched
for can be stated too. Furthermore, certain fold-changes can also
be excluded in a newly introduced field, and groups specified that
should be excluded from the filtering.
This allows to search for fold-change ranges (e.g. 3-6, but
excluding features with higher fold-changes), or to exclude the
fold-changes of groups that are not interesting (e.g. show me
the features with fold-changes >= 3 in groupA/groupB, but don't show
the features with fold-changes >= 3 in groupB/groupA).