BUGFIX: feature selection no longer causes errors
When the blank-sample was selected, clicking on a feature in the
chromatogram resulted in a server error. Some problem about the list not
being able to be serialized to make a json object. This also occured in
all samples of FERMO_session_Soliman.json.
FIX: convert the list to a string and pass that to the client. Then on
clientside the string needs to be parses manually to turn it back into
an array of arrays.

- sampleTable.js and chromatogram.js: do not json.parse the featureTable
- featureTable.js: adjust tableStringtoArray() to account for the table
being passed as a normal string
- route_dashboard.py:
	* convert featureTable to string before delivering the response
    * change ‘response’ to ‘resp’
- view.py: change ‘response’ to ‘resp’
- featureTable.py: correct typehinting and docstring of
update_feature_table()

- chromatogram.js: add try-catch-block to account for empty networks,
for example when it would have been too large (>250 feature in the
network)