Set up the branch for full dockerization
The flask application will go through the WSGI server gunicorn and the
reverse proxy Nginx before it reaches the user. With this commit the
entire branch is set up to allow docker compose to create one container
for the Flask application + gunicorn and one for nginx, which are then
combined.
- Move all files into a folder ‘FERMO’ and create a folder on the same
level as FERMO/ for nginx set up (Dockerfile and config)
- create docker-compose.yml file
- FERMO/Dockerfile:
    - change folder set up to make paths easier
    - use gunicorn to run flask
- requirements.txt: add gunicorn
- move all source files from src/fermo/ up into FERMO/ because otherwise
problems with importing views.py into app.py occur with how the set up
works otherwise (i.e. where the working directory and the Dockerfile are
etc.)