Fix bug: link not working in featureTable
Link to pubchem database did not work as it should, because some
characters of the serach term were removed. The reason was that we
remove certain character-positions from lines in which 'blank' occurs
because these positions are only filled with whitespace which looks bad
if displayed in the table. However, apparently the 'blank' selector was
not specific enough.